fix IndexedBarrierName, and re-enable bounds checking for primitive types

+++ VaultLib.Frameworks.Speed/IndexedBarrierName.cs
@@ -9,7 +9,7 @@ using VaultLib.Core.Types.EA.Reflection;
 namespace VaultLib.Frameworks.Speed
 {
     [VLTTypeInfo(nameof(IndexedBarrierName))]
-    public class IndexedBarrierName : Int32
+    public class IndexedBarrierName : Int16
     {
         public IndexedBarrierName(VltClass @class, VltClassField field, VltCollection collection) : base(@class, field, collection)
         {
